backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
pdfplumber==0.10.3
python-docx==1.1.0
pytesseract==0.3.10
pdfminer.six==20221105
docx2txt==0.8
langchain==0.0.350
langchain-openai==0.0.2
chromadb==0.4.18
pydantic==2.5.0
python-dotenv==1.0.0
aiofiles==23.2.1
typing-extensions==4.8.0
